Replace settings page with backup module
