+++ README.md
@@ -82,7 +82,7 @@ El servidor debe ejecutarse en un único equipo o servidor accesible por la red 
 Todos los navegadores deben por tanto utilizar la misma URL de la API para mantenerse sincronizados.
 
 El archivo activo se guarda en `data/latest.json`.
-Puedes generar copias manualmente con `POST /api/backups` o desde la página **Modo Dev**.
+Puedes generar copias manualmente con `POST /api/backups` o desde la página **Backups**.
 Por defecto se programa una copia diaria en `data/backups/AAAA-MM-DD.json`. Si prefieres desactivar esta tarea define `DISABLE_AUTOBACKUP=1` antes de iniciar el servidor.
 Los respaldos se encuentran en la carpeta `data/backups/`. Cada archivo incluye la base de datos, el historial, la base SQLite y las carpetas de imágenes para que puedas restaurar el estado completo de la aplicación. Si eliminas el repositorio también se borrará esta carpeta a menos que la conserves aparte.
