+++ docs/assets/styles.css
@@ -2679,3 +2679,6 @@ tr:not(.pending) td:first-child::before {
 body.settings-page main {
   animation: fadeIn 0.3s ease-in-out;
 }
+body.backup-page main {
+  animation: fadeIn 0.3s ease-in-out;
+}
