+++ docs/index.html
@@ -47,7 +47,7 @@
   <script type="module" src="js/newClientDialog.js" defer></script>
   <script type="module" src="js/views/home.js" defer></script>
   <script type="module" src="js/views/amfe.js" defer></script>
-  <script type="module" src="js/views/settings.js" defer></script>
+  <script type="module" src="js/views/backup.js" defer></script>
   <script type="module" src="js/views/maestro.js" defer></script>
   <script type="module" src="js/version.js" defer></script>
 </body>
