+++ docs/nav.html
@@ -24,7 +24,7 @@
   <div class="nav-item dropdown">
     <a href="#">Ver más</a>
     <div class="dropdown-menu">
-      <a href="index.html#/settings" class="no-guest">Modo Dev</a>
+      <a href="index.html#/backup" class="no-guest">Backups</a>
       <a href="history.html" class="admin-only">Historial</a>
       <a href="dbviewer.html" class="no-guest">Ver base de datos</a>
     </div>
